refactor(mcp): REFACTOR — share transport fixture across compiled-binary tests
Both test cases open/close a StdioClientTransport + MCP Client + temp
WORKFLOW_STATE_DIR — extract that duplicated plumbing plus the repo-root
discovery and binary-freshness guard into servers/exarchos-mcp/test/process/
_helpers.ts so follow-on integration tests (e.g. task 2.9 bootstrap smoke)
can reuse the same fixture shape without copy-paste.

The helpers module intentionally has no `.test.ts` suffix and starts with
`_` so vitest's glob does not try to collect it as a suite. The leading-
underscore convention is how the MCP server's existing __tests__ signal
"fixtures, not tests" — consistent with the rest of the package.

All assertions, timeouts, and hermeticity contracts unchanged. The test
file drops from 245 lines to 93 and reads as the pure test intent.
